MEAN stack up and running, registration and login set up
Started from scratch due to major errors with backend and front-end.
Angular up and running with a dashboard, home, register and profile.
Express working with mongoDB and angular.
Using JSWebTokens to track users.
Enabled users to only have access to their profile if they are logged
in.
Set up a responsive front-end.
Set up a lot of validation for registration and login, also using bcrypt
to encrpyt passwords.